created balance entity & renamed shared folder into common
