+++ server/common/services/Mailer/templates.js
