+++ server/common/utils/hash.js
