+++ server/components/balances/controller.js
@@ -1,0 +1,4 @@
+export function createBalance(req, res) {
+  // @TODO implement real logic + validation + user checking
+  return res.json({});
+}
