+++ server/components/balances/index.js
@@ -1,0 +1,2 @@
+export { default as BalanceModel } from './model';
+export { default as BalancesRouter } from './routes';
