+++ server/components/balances/routes.js
@@ -1,0 +1,9 @@
+import { Router } from 'express';
+
+import * as BalanceController from './controller';
+
+const route = Router();
+
+route.post('/', BalanceController.createBalance);
+
+export default route;
