+++ server/components/balances/service.js
@@ -1,0 +1,5 @@
+import BalanceModel from './model';
+
+export function createBalance({ name, amount, user }) {
+  return BalanceModel.create({ name, amount, user });
+}
